+++ src/serena/tools/cmd_tools.py
@@ -2,7 +2,7 @@
 Tools supporting the execution of (external) commands
 """
 
-from serena.tools import TOOL_DEFAULT_MAX_ANSWER_LENGTH, Tool, ToolMarkerCanEdit
+from serena.tools import Tool, ToolMarkerCanEdit
 from serena.util.shell import execute_shell_command
 
 
@@ -16,7 +16,7 @@ class ExecuteShellCommandTool(Tool, ToolMarkerCanEdit):
         command: str,
         cwd: str | None = None,
         capture_stderr: bool = True,
-        max_answer_chars: int = TOOL_DEFAULT_MAX_ANSWER_LENGTH,
+        max_answer_chars: int = -1,
     ) -> str:
         """
         Execute a shell command and return its output. If there is a memory about suggested commands, read that first.
@@ -26,7 +26,7 @@ class ExecuteShellCommandTool(Tool, ToolMarkerCanEdit):
         :param cwd: the working directory to execute the command in. If None, the project root will be used.
         :param capture_stderr: whether to capture and return stderr output
         :param max_answer_chars: if the output is longer than this number of characters,
-            no content will be returned. Don't adjust unless there is really no other way to get the content
+            no content will be returned. -1 means using the default value, don't adjust unless there is no other way to get the content
             required for the task.
         :return: a JSON object containing the command's stdout and optionally stderr output
         """